fix: active Flyway (Boot 4), corrige les tests Testcontainers et le schema

+++ src/main/resources/db/migration/V1__schema_initial.sql
@@ -41,7 +41,7 @@ CREATE TABLE type_bonus (
     nom            VARCHAR(50)    UNIQUE NOT NULL,
     categorie      VARCHAR(20)    NOT NULL,
     prix_base      INTEGER        NOT NULL,
-    multiplicateur NUMERIC(10,4)  NOT NULL DEFAULT 0,
+    multiplicateur DOUBLE PRECISION NOT NULL DEFAULT 0,
     description    VARCHAR(200)
 );
